% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg.fit}
\alias{linreg.fit}
\title{Fit a linear regression model using matrix algebra}
\usage{
linreg.fit(x, y)
}
\arguments{
\item{x}{A numeric vector of predictor values}

\item{y}{A numeric vector of response values}
}
\value{
A list with elements including coefficients, residuals, fitted values, and more
}
\description{
Given two vectors, x and y, fit a simple linear regression model using matrix algebra
and return the coefficients, residuals, fitted values, and more.
}
